🚚 [app.routers.system] => split `models` from `router`
